Module opentype::glyph_positioning

The glyph-positioning table.

Structs

Anchor1
An anchor in format 1.

Anchor2
An anchor in format 2.

Anchor3
An anchor in format 3.

Base
A base attachment.

Bases
A set of base attachments.

ChainClassRule
A chaining class positioning rule.

ChainClassRules
A set of chaining class positioning rules.

ChainContextPositioning1
A table for positioning glyphs in a chaining context in format 1.

ChainContextPositioning2
A table for positioning glyphs in a chaining context in format 2.

ChainContextPositioning3
A table for positioning glyphs in a chaining context in format 3.

ChainRule
A chaining positioning rule.

ChainRules
A set of chaining positioning rules.

ClassRule
A class positioning rule.

ClassRules
A set of class positioning rules.

Component
A component attachment.

ContextPositioning1
A table for positioning glyphs in a context in format 1.

ContextPositioning2
A table for positioning glyphs in a context in format 2.

ContextPositioning3
A table for positioning glyphs in a context in format 3.

CursiveAttachment
A table for attaching cursive glyphs.

ExtensionPositioning
A table for other types of positioning.

Ligature
A ligature attachment.

Ligatures
A set of ligature attachments.

Mark1
A mark attachment in format 1.

Mark1s
A set of mark attachments in format 1.

Mark2
A mark attachment in format 2.

Mark2s
A set of mark attachments in format 2.

MarkToBaseAttachment
A table for attaching combining marks to base glyphs.

MarkToLigatureAttachment
A table for attaching combining marks to ligatures.

MarkToMarkAttachment
A table for attaching combining marks to other marks.

Pair1
A pair adjustment in format 1.

Pair1s
A set of pair adjustments in format 1.

Pair2
A pair adjustment in format 2.

Pair2s
A set of pair adjustments in format 2.

PairAdjustment1
A table for adjusting pairs of glyphs in format 1.

PairAdjustment2
A table for adjusting pairs of glyphs in format 2.

Passage
An entry-exit record.

Positioning
A positioning record.

Rule
A positioning rule.

Rules
A set of positioning rules.

Single
A single adjustment.

SingleAdjustment1
A table for adjusting single glyphs in format 1.

SingleAdjustment2
A table for adjusting single glyphs in format 2.

SingleFlags
Single-adjustment flags.

Enums

Anchor
An anchor.

ChainContextPositioning
A table for positioning glyphs in a chaining context.

ContextPositioning
A table for positioning glyphs in a context.

PairAdjustment
A table for adjusting pairs of glyphs.

SingleAdjustment
A table for adjusting single glyphs.

Table
An inner table of a glyph-positioning table.

Type Definitions

GlyphPositioning
A glyph-positioning table.