Module opentype::glyph_substitution

The glyph-substitution table.

Structs

AlternateSubstitution
A table for substituting one glyph with one of many glyphs.

Alternates
A set of alternate substitutions.

ChainClassRule
A chaining class substitution rule.

ChainClassRules
A set of chaining class substitution rules.

ChainContextSubstitution1
A table for substituting glyphs in a chaining context in format 1.

ChainContextSubstitution2
A table for substituting glyphs in a chaining context in format 2.

ChainContextSubstitution3
A table for substituting glyphs in a chaining context in format 3.

ChainRule
A chaining substitution rule.

ChainRules
A set of chaining substitution rules.

ClassRule
A class substitution rule.

ClassRules
A set of class substitution rules.

ContextSubstitution1
A table for substituting glyphs in a context in format 1.

ContextSubstitution2
A table for substituting glyphs in a context in format 2.

ContextSubstitution3
A table for substituting glyphs in a context in format 3.

ExtensionSubstitution
A table for other types of substitution.

Ligature
A ligature substitution.

LigatureSubstitution
A table for substituting multiple glyphs with one glyph.

Ligatures
A set of ligature substitutions.

MultipleSubstitution
A table for substituting one glyph with more than one glyph.

ReverseChainContextSubstitution
A table for substituting glyphs in reverse order in a chaining context.

Rule
A substitution rule.

Rules
A set of substitution rules.

Sequence
A substitution sequence of glyphs.

SingleSubstitution1
A table for substituting one glyph with one glyph in format 1.

SingleSubstitution2
A table for substituting one glyph with one glyph in format 2.

Substitution
A substitution operation.

Enums

ChainContextSubstitution
A table for substituting glyphs in a chaining context.

ContextSubstitution
A table for substituting glyphs in a context.

SingleSubstitution
A table for substituting one glyph with one glyph.

Table
An inner table of a glyph-substitution table.

Type Definitions

GlyphSubstitution
A glyph-substitution table.